+++ cmake/tests.cmake
@@ -57,6 +57,8 @@ add_executable(CITests
   ##tests/qcstack/QuantumBrillianceRemoteAcceleratorTester.cpp # unit-test harness for Qcstack server - Qristal
   tests/placement/tketTester.cpp
   tests/transpilationTester.cpp
+  tests/qobj/qobjTester.cpp
+  tests/QuantumBrillianceAcceleratorTester.cpp
 )
 
 add_test(NAME ci_tester COMMAND CITests)