+++ include/qb/core/session.hpp
@@ -1427,7 +1427,7 @@ namespace qb
       /// Helper to populate result tables (e.g. counts, expectation values, resource estimations) post-execution.
       void process_run_result(const std::size_t ii, const std::size_t jj, const run_i_j_config& run_config,
                               std::shared_ptr<xacc::AcceleratorBuffer> buffer_b, double runtime_ms,
-                              std::shared_ptr<xacc::quantum::QuantumBrillianceAccelerator> qb_transpiler);
+                              std::shared_ptr<qb::QuantumBrillianceAccelerator> qb_transpiler);
       /// Util method to compile input source string into IR
       /// This method is thread-safe, thus can be used to compile multiple source strings in parallel.
       std::shared_ptr<xacc::CompositeInstruction> compile_input(const std::string& in_source_string, int in_num_qubits, source_string_type in_source_type);