+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Qristal is a full-stack SDK for quantum accelerators.
 
 - Transpilation: wrapping of gate angles to [-pi,pi] for native gates
 
+- Transpilation: out_transpiled_circuit (OpenQASM 2.0) in native gates.
+
+- Profiler: counting number of gates and timing data after transpilation to the native gate set.
+
 ### Fixed
 
 - Minor documentation fixes